requirements.txt
google-cloud-storage==2.14.0
pandas==2.1.4
google-cloud-firestore==2.14.0
# Prefer google-genai (new); falls back to google-generativeai if not installed
google-genai>=1.0.0
google-generativeai>=0.3.2
tqdm==4.66.1
python-dotenv==1.0.0